requirements.txt
Django==3.2.5
django-ckeditor==6.2.0
django-cleanup==6.0.0
django-crispy-forms==1.14.0
django-js-asset==1.0.0